Kubespray 2.19

+++ 3_NFS_K8s_dashboard.sh
@@ -8,11 +8,18 @@ ORANGE='\033[0;33m'
 BLUE='\033[0;34m'
 NC='\033[0m'
 
-NFS_SERVER=$(hostname -i)
+echo -e "Creating docker secret"
+kubectl create secret generic regcred --from-file=.dockerconfigjson=/root/.docker/config.json --type=kubernetes.io/dockerconfigjson
 
 echo -e "Installing NFS provisioner"
 
-wget https://get.helm.sh/helm-v3.7.2-linux-ppc64le.tar.gz && tar -xvf helm* && mv linux-ppc64le/helm /usr/local/bin/helm && rm -rf helm-v* linux-ppc64le
+NFS_SERVER=$(hostname -i)
+
+if ! command -v helm &> /dev/null
+then
+    wget https://get.helm.sh/helm-v3.7.2-linux-ppc64le.tar.gz && tar -xvf helm* && mv linux-ppc64le/helm /usr/local/bin/helm && rm -rf helm-v* linux-ppc64le
+fi
+
 helm repo add nfs-subdir-external-provisioner https://kubernetes-sigs.github.io/nfs-subdir-external-provisioner/
 helm install nfs-subdir-external-provisioner nfs-subdir-external-provisioner/nfs-subdir-external-provisioner \
     --set nfs.server=${NFS_SERVER} \
